split out search module
